requirements.txt
Flask==2.3.2
Pillow==10.0.0
rembg
python-dotenv==1.0.0
uuid==1.30